# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Goat Screams API
  version: 1.0.0
  description: |
    Public REST API for browsing curated goat screams, powered by MongoDB and Cloudinary.
    All endpoints are rate-limited. Moderation routes require the `x-admin-token` header.

    ## API Versioning
    All API endpoints are versioned under `/api/v1/`. Legacy `/api/*` requests are redirected
    to the corresponding `/api/v1/*` endpoint with a 308 Permanent Redirect.

    ## Filter Composition Rules

    Filters can be combined to narrow down results. All filters use **AND** logic — each
    additional filter further restricts the result set.

    ### List Endpoints (`/api/v1/screams`, `/api/v1/screams/random`)

    | Filter | Type | Description | Example |
    |--------|------|-------------|---------|
    | `intensity_min` | integer | Minimum intensity (1-10) | `?intensity_min=7` |
    | `intensity_max` | integer | Maximum intensity (1-10) | `?intensity_max=9` |
    | `year` | integer | Exact year match | `?year=2020` |
    | `source_type` | string | Source category | `?source_type=viral_video` |
    | `meme_status` | string | Meme lifecycle status | `?meme_status=classic` |
    | `breed` | string | Goat breed (case-insensitive) | `?breed=Alpine` |
    | `category` | string | Audio category | `?category=short_burst` |

    **Combining filters:**
    ```
    GET /api/v1/screams?intensity_min=7&intensity_max=10&year=2020
    ```
    Returns screams with intensity 7-10 **AND** from year 2020.

...
servers:
...
components:
  securitySchemes:
    AdminToken:
      type: apiKey
      in: header
      name: x-admin-token
  responses:
    ValidationError:
      description: Request failed validation
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    Unauthorized:
      description: Authentication or authorization failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    RateLimited:
      description: Too many requests; retry after the window resets
      headers:
        Retry-After:
          description: Seconds until a new request is allowed
          schema:
            type: integer
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    ServerError:
      description: Unexpected server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
  parameters:
    Page:
      in: query
      name: page
      description: Page number
      schema:
        type: integer
        default: 1
        minimum: 1
    Limit:
      in: query
      name: limit
      description: Page size
      schema:
        type: integer
        default: 20
        minimum: 1
        maximum: 100
    Status:
      in: query
      name: status
      description: Filter by moderation status
      schema:
        type: string
        enum:
          - pending_review
          - approved
          - rejected
          - all
...
  schemas:
    ApiError:
      type: object
      description: Standard error envelope returned by the API
      properties:
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
              description: Stable error code (e.g., VALIDATION_ERROR)
            message:
              type: string
              description: Human-readable message for the error
            request_id:
              type: string
              description: Request correlation identifier
            details:
              description: Additional context, often validation failures
              oneOf:
                - type: object
                  items:
                    type: object
      required:
        - error
...
    Submission:
      type: object
      description: A user-submitted goat scream awaiting moderation
      properties:
        id:
          type: string
          description: Submission identifier
          pattern: '^sub_[0-9a-fA-F-]{36}$'
        title:
          type: string
          description: Submitted title
        status:
          type: string
          description: Moderation status
          enum:
            - pending_review
            - approved
            - rejected
        source:
          type: object
          description: Optional source metadata provided by the submitter
          properties:
            title:
              type: string
            url:
              type: string
              format: uri
            platform:
              type: string
            author:
              type: string
        context:
          type: string
          description: Additional context about the clip
        year:
          type: integer
          description: Year associated with the clip
          minimum: 1900
          maximum: 2100
        analysis:
          type: object
          description: Optional audio analysis metadata
        audio:
          type: object
          description: Audio metadata
          properties:
            cloudinary_url:
              type: string
              format: uri
              description: Location of the uploaded audio asset
            duration:
              type: number
              description: Duration in seconds
              minimum: 0
            intensity:
              type: integer
              description: Relative intensity from 1-10
              minimum: 1
              maximum: 10
            category:
              type: string
              description: Audio category label
        goat_scream_id:
          type: string
          description: Linked scream id when promoted
          pattern: '^[A-Za-z0-9._-]+'
        review_notes:
          type: string
          description: Moderator notes
        createdAt:
          type: string
          format: date-time
          description: Creation timestamp
        updatedAt:
          type: string
          format: date-time
          description: Last update timestamp
    SubmissionPage:
      type: object
      description: Paginated submissions
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        items:
          type: array
          items:
            $ref: '#/components/schemas/Submission'
      required:
        - page
        - limit
        - total
        - items
    SubmissionAccepted:
      type: object
      description: Accepted submission payload
      properties:
        message:
          type: string
        estimatedReview:
          type: string
          description: Human-readable review estimate
        submission:
          type: object
          properties:
            id:
              type: string
            status:
              type: string
            analysis:
              type: object
            audio_url:
              type: string
              format: uri
      required:
        - message
        - submission
...
paths:
...
  /api/v1/submissions:
    post:
      operationId: submitScream
      summary: Submit a new goat scream
      description: Accepts multipart or JSON submissions with `title` and audio, returning queued submission info (`message`, `estimatedReview`, `submission`).
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - title
              properties:
                title:
                  type: string
                  description: Title for the submission
                  minLength: 3
                context:
                  type: string
                  description: Background or story for the clip
                year:
                  type: integer
                  description: Year associated with the clip
                  minimum: 1900
                  maximum: 2100
                audio:
                  type: string
                  format: binary
                  description: Audio file upload
                tags:
                  type: string
                  description: Comma-separated tags
          application/json:
            schema:
              type: object
              required:
                - title
                - audio_url
              properties:
                title:
                  type: string
                  description: Title for the submission
                  minLength: 3
                audio_url:
                  type: string
                  format: uri
                  description: Remote audio URL to import
                context:
                  type: string
                year:
                  type: integer
                  minimum: 1900
                  maximum: 2100
      responses:
        '202':
          description: Submission queued for review
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubmissionAccepted'
        '400':
          $ref: '#/components/responses/ValidationError'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/ServerError'
    get:
      operationId: listSubmissions
      summary: List submissions
      description: Returns paginated submissions filtered by `status` with `page`/`limit` controls; response includes `page`, `limit`, `total`, and `items` array of `Submission`.
      parameters:
        - $ref: '#/components/parameters/Status'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Submission queue
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubmissionPage'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/ServerError'
  /api/v1/moderation/submissions:
    get:
      operationId: listSubmissionsForModeration
      summary: List submissions for moderation
      description: Moderation-only listing filtered by `status` with `page`/`limit` controls; response includes `page`, `limit`, `total`, and `items` array of `Submission`.
      security:
        - AdminToken: []
      parameters:
        - $ref: '#/components/parameters/Status'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Submission list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubmissionPage'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/ServerError'
  /api/v1/moderation/submissions/{id}/approve:
    patch:
      operationId: approveSubmission
      summary: Approve and promote a submission
      description: Promotes a submission identified by `id`, optionally overriding `meme_status`, `tags`, `license_*`, `title`, `source_type`, and `goat_scream_id`; responds with `message`, `goat_scream_id`, and `media_url`.
      security:
        - AdminToken: []
      parameters:
        - in: path
          name: id
          required: true
          description: Submission identifier
          schema:
            type: string
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                meme_status:
                  type: string
                tags:
                  type: string
                  description: Comma-separated tags to apply
                license_type:
                  type: string
                license_url:
                  type: string
                  format: uri
                attribution_required:
                  type: boolean
                attribution_text:
                  type: string
                review_notes:
                  type: string
                goat_scream_id:
                  type: string
                source_type:
                  type: string
                title:
                  type: string
      responses:
        '200':
          description: Submission approved
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  goat_scream_id:
                    type: string
                  media_url:
                    type: string
                    format: uri
                required:
                  - message
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/ServerError'
  /api/v1/moderation/submissions/{id}/reject:
    patch:
      operationId: rejectSubmission
      summary: Reject a submission
      description: Rejects a submission identified by `id`, with optional `review_notes` and `delete_media` toggle; responds with `message` and `id`.
      security:
        - AdminToken: []
      parameters:
        - in: path
          name: id
          required: true
          description: Submission identifier
          schema:
            type: string
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                review_notes:
                  type: string
                delete_media:
                  type: boolean
      responses:
        '200':
          description: Submission rejected
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  id:
                    type: string
                required:
                  - message
                  - id
        '400':
          $ref: '#/components/responses/ValidationError'
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  id:
                    type: string
                required:
                  - message
                  - id
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/ServerError'